+++ tergite_autocalibration/lib/nodes/characterization/purity_benchmarking/tests/test_analysis.py
@@ -33,7 +33,9 @@ class TestPurityBenchmarkingAnalysis(unittest.TestCase):
     def setUpClass(self):
         # Setup the dataset for testing from a file
         test_dir = Path(__file__).parent
-        file_path = test_dir / "data" / "dataset_0.hdf5"  # "testdata.hdf5"
+        file_path = (
+            test_dir / "data" / "dataset_purity_benchmarking_0.hdf5"
+        )  # "testdata.hdf5"
         self.dataset = xr.open_dataset(file_path)
 
     @unittest.skip
@@ -47,7 +49,6 @@ class TestPurityBenchmarkingAnalysis(unittest.TestCase):
             self.analysis.number_of_repetitions, self.dataset.sizes.get("seed", 1)
         )
 
-    @unittest.skip
     def test_run_fitting(self):
         analysis = PurityBenchmarkingQubitAnalysis("name", ["redis_field"])
         analysis.process_qubit(self.dataset, "yq06")
@@ -68,7 +69,6 @@ class TestPurityBenchmarkingAnalysis(unittest.TestCase):
         self.assertTrue(0 <= fidelity[0] <= 1.002)
         self.assertIsInstance(analysis.fit_results, ModelResult)
 
-    @unittest.skip
     def test_plotter(self):
         analysis = PurityBenchmarkingQubitAnalysis("name", ["redis_field"])
         analysis.process_qubit(self.dataset, "yq14")