+++ tergite_autocalibration/lib/nodes/coupler/cz_parametrisation/analysis.py
@@ -349,7 +349,7 @@ class CZParametrisationFixDurationCouplerAnalysis(BaseCouplerAnalysis):
                 pass
 
 
-class CZParametrisationFixDurationNodeAnalysis(BaseAllCouplersRepeatAnalysis):
+class CZParametrizationFixDurationNodeAnalysis(BaseAllCouplersRepeatAnalysis):
     single_coupler_analysis_obj = CZParametrisationFixDurationCouplerAnalysis
 
     def __init__(self, name, redis_fields):