+++ tergite_autocalibration/lib/nodes/coupler/cz_parametrisation/tests/test_coupler_analysis.py
@@ -18,16 +18,11 @@ import pytest
 import xarray as xr
 
 from tergite_autocalibration.lib.base.analysis import (
-    BaseAllCouplersAnalysis,
-    BaseAllCouplersRepeatAnalysis,
-    BaseAllQubitsRepeatAnalysis,
     BaseAnalysis,
     BaseCouplerAnalysis,
-    BaseNodeAnalysis,
 )
 from tergite_autocalibration.lib.nodes.coupler.cz_parametrisation.analysis import (
     CZParametrisationFixDurationCouplerAnalysis,
-    CZParametrisationFixDurationNodeAnalysis,
     CombinedFrequencyVsAmplitudeAnalysis,
     FrequencyVsAmplitudeQ1Analysis,
     FrequencyVsAmplitudeQ2Analysis,
@@ -189,13 +184,17 @@ def test_PickGoodValueIfSmallestInAbsolute(
 @pytest.fixture(autouse=True)
 def setup_data_mutliple_files():
     # It should be a single dataset, but we do not have one yet, so we loop over existing files
-    dataset_path = Path(__file__).parent / "data" / "dataset_0.hdf5"
+    dataset_path = (
+        Path(__file__).parent
+        / "data"
+        / "dataset_cz_parametrization_fix_duration_0.hdf5"
+    )
     ds = xr.open_dataset(dataset_path, engine="scipy")
     combined_dataset = ds
 
     # combined_dataset = xr.Dataset()
     for i in (1, 2, 3):
-        filename = "dataset_" + str(i) + ".hdf5"
+        filename = "dataset_cz_parametrization_fix_duration_" + str(i) + ".hdf5"
         dataset_path = Path(__file__).parent / "data" / filename
         ds = xr.open_dataset(dataset_path, engine="scipy")
         combined_dataset = xr.merge([combined_dataset, ds])
@@ -208,7 +207,6 @@ def setup_data_mutliple_files():
     return combined_dataset, freqs, amps
 
 
-@unittest.skip
 def test_PickLowestCurrentCompleteAnalysis(
     setup_data_mutliple_files: tuple[xr.Dataset, ndarray, ndarray]
 ):